refactor(server): Implement gap buffer positioning for columns and cards
Replace integer sequencing (1, 2, 3) with gap buffer technique using multiples of 100.
This eliminates UNIQUE constraint conflicts during reordering, simplifies position
management, and removes the need for offset workarounds. Positions now initialize
at 100, 200, 300... for improved flexibility in future enhancements.

Changes:
- schema.sql: Change position columns from INTEGER to BIGINT
- columns.js: Initialize new columns at (count+1)*100, reorder without offset
- cards.js: Initialize new cards at (count+1)*100, resequence with gap buffer

+++ server/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE columns (
   id SERIAL PRIMARY KEY,
   board_id INTEGER NOT NULL REFERENCES boards(id) ON DELETE CASCADE,
   name VARCHAR(120) NOT NULL,
-  position INTEGER NOT NULL,
+  position BIGINT NOT NULL,
   created_at TIMESTAMP NOT NULL DEFAULT NOW(),
   UNIQUE (board_id, position)
 );
@@ -19,7 +19,7 @@ CREATE TABLE cards (
   title VARCHAR(200) NOT NULL,
   description TEXT,
   due_date DATE,
-  position INTEGER NOT NULL,
+  position BIGINT NOT NULL,
   created_at TIMESTAMP NOT NULL DEFAULT NOW(),
   UNIQUE (column_id, position)
 );
